Move scene mesh tooling under MuJoCo
