+++ dimos/simulation/mujoco/mesh_scene.py
